java_analyzer_py/requirements.txt
javalang
flask
flaskwebgui 